django_server/MythicTable/requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonpointer==2.3
pip==20.3.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
pyOpenSSL==23.1.1
pytz==2023.3
requests==2.29.0
rest-framework-simplejwt==0.0.2
service-identity==21.1.0
setuptools==52.0.0
six==1.16.0
slugify==0.0.1
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
urllib3==1.26.15
wheel==0.34.2
zope.interface==6.0